+++ CHANGELOG.md
@@ -1,14 +1,15 @@
-# Change Log
+# Changelog
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/)
-and this project adheres to [Semantic Versioning](http://semver.org/).
+The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
+and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 ### Added
 - Added chan as default export to expose api (init, changes & release)
 
 ### Changed
+- Updated Changelog head wording according to [Keep a Changelog](http://keepachangelog.com/) v1.0.0
 - Readme with better documentation for API and programmatic usage
 
 ## [1.3.0-0] - 2017-02-10 [YANKED]